refactor(components): extract CameraScanner, ManualEntry, CredentialCard; add useCameraPermission hook; centralize Credential type; add components README; update screens to use components
