+++ expoapp/app/credential-list.tsx
@@ -5,52 +5,8 @@ import { FontAwesome } from '@expo/vector-icons';
 import { useRouter, Stack } from 'expo-router';
 
 import useCredentials from '../hooks/useCredentials';
-import { generateTotp } from '../utils/generateTotp';
-
-export type Credential = {
-  accountName: string;
-  issuer?: string;
-  secret: string;
-  _key?: string;
-};
-
-function CredentialCard({ credential, onDelete }: { credential: Credential; onDelete: () => void }) {
-  const [code, setCode] = useState('');
-  const [timeLeft, setTimeLeft] = useState(30 - (Math.floor(Date.now() / 1000) % 30));
-
-  useEffect(() => {
-    function updateCode() {
-      setCode(generateTotp(credential.secret));
-      setTimeLeft(30 - (Math.floor(Date.now() / 1000) % 30));
-    }
-    updateCode();
-    const interval = setInterval(updateCode, 1000);
-    return () => clearInterval(interval);
-  }, [credential.secret]);
-
-  return (
-    <View style={{ padding: 12, borderBottomWidth: 1, borderColor: '#eee', flexDirection: 'row', alignItems: 'center', justifyContent: 'space-between' }}>
-      <View>
-        <Text style={{ fontWeight: 'bold', fontSize: 16 }}>{credential.accountName}</Text>
-        <Text>Issuer: {credential.issuer || 'N/A'}</Text>
-        <Text>Code: <Text style={{ fontFamily: 'monospace', fontSize: 18 }}>{code}</Text> <Text style={{ color: '#888' }}>({timeLeft}s)</Text></Text>
-      </View>
-      <Pressable
-        onPress={onDelete}
-        accessibilityLabel={`Delete ${credential.accountName}`}
-        style={({ pressed }) => [{
-          padding: 8,
-          borderRadius: 20,
-          backgroundColor: pressed ? '#eee' : 'transparent',
-          alignItems: 'center',
-          justifyContent: 'center',
-        }]}
-      >
-        <FontAwesome name="trash" size={20} color="#888" />
-      </Pressable>
-    </View>
-  );
-}
+import { Credential } from '../types/credential';
+import CredentialCard from '../components/CredentialCard';
 
 export default function CredentialListScreen() {
   const { credentials, loading, remove, reload } = useCredentials();
