+++ expoapp/types/credential.ts
@@ -1,0 +1,6 @@
+export type Credential = {
+  accountName: string;
+  issuer?: string;
+  secret: string;
+  _key?: string;
+};
